test: replace type-checking with tsd notation
